requirements.txt
appnope==0.1.0
certifi==2017.7.27.1
chardet==3.0.4
decorator==4.1.2
defusedxml==0.5.0
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.6
django-anymail==1.0
django-mptt==0.8.7
django-simple-email-confirmation==0.21
django-widget-tweaks==1.4.1
djangorestframework==3.7.1
gunicorn==19.4.1
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.0
model-mommy==1.4.0
oauthlib==2.0.5
olefile==0.44
parso==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.3.0
prompt-toolkit==1.0.15
psycopg2==2.6.1
ptyprocess==0.5.2
Pygments==2.2.0
PyJWT==1.5.3
python-decouple==3.1
python3-openid==3.1.0
pytz==2017.2
raven==6.2.1
requests==2.18.4
requests-oauthlib==0.8.0
simplegeneric==0.8.1
six==1.11.0
social-auth-app-django==1.2.0
social-auth-core==1.4.0
static3==0.7.0
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
